+++ src/anemoi/graphs/generate/hex_icosahedron.py
@@ -1,3 +1,12 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
 from __future__ import annotations
 
 import h3
@@ -40,8 +49,8 @@ def create_hex_nodes(
     node_ordering = get_coordinates_ordering(coords_rad)
 
     if area_mask_builder is not None:
-        aoi_mask = area_mask_builder.get_mask(coords_rad)
-        node_ordering = node_ordering[aoi_mask[node_ordering]]
+        area_mask = area_mask_builder.get_mask(coords_rad)
+        node_ordering = node_ordering[area_mask[node_ordering]]
 
     graph = create_nx_graph_from_hex_coords(nodes, node_ordering)