+++ src/anemoi/graphs/generate/tri_icosahedron.py
@@ -1,3 +1,12 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
 from __future__ import annotations
 
 from collections.abc import Iterable
@@ -13,7 +22,7 @@ from anemoi.graphs.generate.utils import get_coordinates_ordering
 
 
 def create_tri_nodes(
-    resolution: int, aoi_mask_builder: KNNAreaMaskBuilder | None = None
+    resolution: int, area_mask_builder: KNNAreaMaskBuilder | None = None
 ) -> tuple[nx.DiGraph, np.ndarray, list[int]]:
     """Creates a global mesh from a refined icosahedron.
 
@@ -23,7 +32,7 @@ def create_tri_nodes(
     ----------
     resolution : int
         Level of mesh resolution to consider.
-    aoi_mask_builder : KNNAreaMaskBuilder
+    area_mask_builder : KNNAreaMaskBuilder
         KNNAreaMaskBuilder with the cloud of points to limit the mesh area, by default None.
 
     Returns
@@ -35,20 +44,83 @@ def create_tri_nodes(
     node_ordering : list[int]
         Order of the node coordinates to be sorted by latitude and longitude.
     """
-    sphere = trimesh.creation.icosphere(subdivisions=resolution, radius=1.0)
-
-    coords_rad = cartesian_to_latlon_rad(sphere.vertices)
+    coords_rad = get_latlon_coords_icosphere(resolution)
 
     node_ordering = get_coordinates_ordering(coords_rad)
 
-    if aoi_mask_builder is not None:
-        aoi_mask = aoi_mask_builder.get_mask(coords_rad)
-        node_ordering = node_ordering[aoi_mask[node_ordering]]
+    if area_mask_builder is not None:
+        area_mask = area_mask_builder.get_mask(coords_rad)
+        node_ordering = node_ordering[area_mask[node_ordering]]
 
     # Creates the graph, with the nodes sorted by latitude and longitude.
     nx_graph = create_nx_graph_from_tri_coords(coords_rad, node_ordering)
 
     return nx_graph, coords_rad, list(node_ordering)
+
+
+def create_stretched_tri_nodes(
+    base_resolution: int,
+    lam_resolution: int,
+    area_mask_builder: KNNAreaMaskBuilder | None = None,
+) -> tuple[nx.DiGraph, np.ndarray, list[int]]:
+    """Creates a global mesh with 2 levels of resolution.
+
+    The base resolution is used to define the nodes outside the Area Of Interest (AOI),
+    while the lam_resolution is used to define the nodes inside the AOI.
+
+    Parameters
+    ---------
+    base_resolution : int
+        Global resolution level.
+    lam_resolution : int
+        Local resolution level.
+    area_mask_builder : KNNAreaMaskBuilder
+        NearestNeighbors with the cloud of points to limit the mesh area.
+
+    Returns
+    -------
+    nx_graph : nx.DiGraph
+        The graph with the added nodes.
+    coords_rad : np.ndarray
+        The node coordinates (not ordered) in radians.
+    node_ordering : list[int]
+        Order of the node coordinates to be sorted by latitude and longitude.
+    """
+    assert area_mask_builder is not None, "AOI mask builder must be provided to build refined grid."
+    # Get the low resolution nodes outside the AOI
+    base_coords_rad = get_latlon_coords_icosphere(base_resolution)
+    base_area_mask = ~area_mask_builder.get_mask(base_coords_rad)
+
+    # Get the high resolution nodes inside the AOI
+    lam_coords_rad = get_latlon_coords_icosphere(lam_resolution)
+    lam_area_mask = area_mask_builder.get_mask(lam_coords_rad)
+
+    coords_rad = np.concatenate([base_coords_rad[base_area_mask], lam_coords_rad[lam_area_mask]])
+
+    node_ordering = get_coordinates_ordering(coords_rad)
+
+    # Creates the graph, with the nodes sorted by latitude and longitude.
+    nx_graph = create_nx_graph_from_tri_coords(coords_rad, node_ordering)
+
+    return nx_graph, coords_rad, list(node_ordering)
+
+
+def get_latlon_coords_icosphere(resolution: int) -> np.ndarray:
+    """Get the latitude and longitude coordinates (in radians) of the icosphere.
+
+    Parameters
+    ----------
+    resolution : int
+        The resolution level of the icosphere.
+
+    Returns
+    -------
+    np.ndarray
+        The latitude and longitude coordinates, in radians, of the icosphere.
+    """
+    sphere = trimesh.creation.icosphere(subdivisions=resolution, radius=1.0)
+    coords_rad = cartesian_to_latlon_rad(sphere.vertices)
+    return coords_rad
 
 
 def create_nx_graph_from_tri_coords(coords_rad: np.ndarray, node_ordering: np.ndarray) -> nx.DiGraph:
@@ -92,26 +164,18 @@ def add_edges_to_nx_graph(
     """
     assert x_hops > 0, "x_hops == 0, graph would have no edges ..."
 
-    sphere = trimesh.creation.icosphere(subdivisions=resolutions[-1], radius=1.0)
-    vertices_rad = cartesian_to_latlon_rad(sphere.vertices)
-    node_neighbours = get_neighbours_within_hops(sphere, x_hops, valid_nodes=list(graph.nodes))
-
-    for idx_node, idx_neighbours in node_neighbours.items():
-        add_neigbours_edges(graph, idx_node, idx_neighbours)
-
-    tree = BallTree(vertices_rad, metric="haversine")
+    graph_vertices = np.array([graph.nodes[i]["hcoords_rad"] for i in sorted(graph.nodes)])
+    tree = BallTree(graph_vertices, metric="haversine")
 
     # Build the multi-scale connections
-    for resolution in resolutions[:-1]:
-        # Define the isophere at specified 'resolution' level
+    for resolution in resolutions:
+        # Define the coordinates of the isophere vertices at specified 'resolution' level
         r_sphere = trimesh.creation.icosphere(subdivisions=resolution, radius=1.0)
-
-        # Get the vertices of the isophere
         r_vertices_rad = cartesian_to_latlon_rad(r_sphere.vertices)
 
         # Limit area of mesh points.
         if area_mask_builder is not None:
-            aoi_mask = area_mask_builder.get_mask(vertices_rad)
+            aoi_mask = area_mask_builder.get_mask(r_vertices_rad)
             valid_nodes = np.where(aoi_mask)[0]
         else:
             valid_nodes = None
@@ -120,7 +184,7 @@ def add_edges_to_nx_graph(
 
         _, vertex_mapping_index = tree.query(r_vertices_rad, k=1)
         for idx_node, idx_neighbours in node_neighbours.items():
-            add_neigbours_edges(graph, idx_node, idx_neighbours, vertex_mapping_index=vertex_mapping_index)
+            graph = add_neigbours_edges(graph, idx_node, idx_neighbours, vertex_mapping_index=vertex_mapping_index)
 
     return graph
 
@@ -137,7 +201,8 @@ def get_neighbours_within_hops(
     x_hops : int
         Number of hops between 2 nodes to consider them neighbours.
     valid_nodes : list[int], optional
-        The list of valid nodes to consider, by default None.
+        List of valid nodes to consider, by default None. It is useful to consider only a subset of the nodes to save
+        computation time.
 
     Returns
     -------
@@ -166,7 +231,7 @@ def add_neigbours_edges(
     neighbour_indices: Iterable[int],
     self_loops: bool = False,
     vertex_mapping_index: np.ndarray | None = None,
-) -> None:
+) -> nx.Graph:
     """Adds the edges of one node to its neighbours.
 
     Parameters
@@ -181,18 +246,27 @@ def add_neigbours_edges(
         Whether is supported to add self-loops, by default False.
     vertex_mapping_index : np.ndarray, optional
         Index to map the vertices from the refined sphere to the original one, by default None.
+
+    Returns
+    -------
+    nx.Graph
+        The graph with the added edges.
     """
+    graph_nodes_idx = list(sorted(graph.nodes))
     for neighbour_idx in neighbour_indices:
         if not self_loops and node_idx == neighbour_idx:  # no self-loops
             continue
 
         if vertex_mapping_index is not None:
             # Use the same method to add edge in all spheres
-            node_neighbour = vertex_mapping_index[neighbour_idx][0]
-            node = vertex_mapping_index[node_idx][0]
+            node_neighbour = graph_nodes_idx[vertex_mapping_index[neighbour_idx][0]]
+            node = graph_nodes_idx[vertex_mapping_index[node_idx][0]]
         else:
-            node, node_neighbour = node_idx, neighbour_idx
+            node_neighbour = graph_nodes_idx[neighbour_idx]
+            node = graph_nodes_idx[node_idx]
 
         # add edge to the graph
         if node in graph and node_neighbour in graph:
             graph.add_edge(node_neighbour, node)
+
+    return graph