+++ src/anemoi/graphs/nodes/builders/base.py
@@ -1,3 +1,12 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
 from __future__ import annotations
 
 from abc import ABC
@@ -25,15 +34,20 @@ class BaseNodeBuilder(ABC):
 
     def __init__(self, name: str) -> None:
         self.name = name
-        self.aoi_mask_builder = None
+        self.area_mask_builder = None
 
-    def register_nodes(self, graph: HeteroData) -> None:
+    def register_nodes(self, graph: HeteroData) -> HeteroData:
         """Register nodes in the graph.
 
         Parameters
         ----------
         graph : HeteroData
             The graph to register the nodes.
+
+        Returns
+        -------
+        HeteroData
+            The graph with the registered nodes.
         """
         graph[self.name].x = self.get_coordinates()
         graph[self.name].node_type = type(self).__name__