+++ CHANGELOG.md
@@ -11,14 +11,31 @@ Keep it human-readable, your future self will thank you!
 ## [Unreleased](https://github.com/ecmwf/anemoi-graphs/compare/0.3.0...HEAD)
 
 ### Added
+
+- fix: bug in color when plotting isolated nodes
+
+### Added
+
+- Add anemoi-transform link to documentation
+
+### Added
 - ci: hpc-config, CODEOWNERS (#49)
 - feat: New node builder class, CutOutZarrDatasetNodes, to create nodes from 2 datasets. (#30)
 - feat: New class, KNNAreaMaskBuilder, to specify Area of Interest (AOI) based on a set of nodes. (#30)
 - feat: New node builder classes, LimitedAreaXXXXXNodes, to create nodes within an Area of Interest (AOI). (#30)
 - feat: Expanded MultiScaleEdges to support multi-scale connections in limited area graphs. (#30)
+- feat: New method update_graph(graph) in the GraphCreator class. (#60)
+- feat: New class StretchedTriNodes to create a stretched mesh. (#51)
+- feat: Expanded MultiScaleEdges to support multi-scale connections in stretched graphs. (#51)
+- feat: Add CONTRIBUTORS.md file. (#72)
 
 ### Changed
 - ci: small fixes and updates pre-commit, downsteam-ci (#49)
+- Update CODEOWNERS
+- Fix pre-commit regex
+- ci: extened python versions to include 3.11 and 3.12
+- Update copyright notice
+- Fix `__version__` import in init
 
 ## [0.3.0 Anemoi-graphs, minor release](https://github.com/ecmwf/anemoi-graphs/compare/0.2.1...0.3.0) - 2024-09-03